backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
geoalchemy2==0.9.4
python-dotenv==0.19.0
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
pydantic==1.8.2
email-validator==1.1.3
ortools==9.0.9048
requests==2.26.0
shapely==1.7.1
geopy==2.2.0
python-multipart==0.0.5
async-exit-stack==1.0.1
async-generator==1.10
